Struct ros2_logger::Ros2Logger

pub struct Ros2Logger { /* private fields */ }

Implements Log and a set of simple builder methods for configuration.

Use the various “builder” methods on this struct to configure the logger, then call init to configure the log crate.

Implementations

impl Ros2Logger

pub fn new(node: Arc<Node>) -> Ros2Logger

Initializes the global logger with a Ros2Logger instance with default log level set to Level::Trace.

use std::env;
use std::sync::Arc;
use ros2_logger::Ros2Logger;

let context = rclrs::Context::new(env::args())?;
let node = rclrs::create_node(&context, "minimal_node")?;
Ros2Logger::new(Arc::clone(&node)).env().init().unwrap();
log::warn!("This is an example message.");

pub fn env(self) -> Ros2Logger

Enables the user to choose log level by setting RUST_LOG=<level> environment variable. This will use the default level set by with_level if RUST_LOG is not set or can’t be parsed as a standard log level.

This must be called after with_level. If called before with_level, it will have no effect.

pub fn with_level(self, level: LevelFilter) -> Ros2Logger

Set the ‘default’ log level.

You can override the default level for specific modules and their submodules using with_module_level

This must be called before env. If called after env, it will override the value loaded from the environment.

pub fn with_module_level(self, target: &str, level: LevelFilter) -> Ros2Logger

Override the log level for some specific modules.

This sets the log level of a specific module and all its submodules. When both the level for a parent module and a child module are set, the more specific value is taken. If the log level for the same module is specified twice, the resulting log level is implementation defined.

Examples

Silence an overly verbose crate:

use std::sync::Arc;
use ros2_logger::Ros2Logger;
use log::LevelFilter;

let context = rclrs::Context::new(std::env::args())?;
let node = Arc::new(rclrs::Node::new(&context, "test")?);
Ros2Logger::new(Arc::clone(&node)).with_module_level("chatty_dependency", LevelFilter::Warn).init().unwrap();

Disable logging for all dependencies:

use std::sync::Arc;
use ros2_logger::Ros2Logger;
use log::LevelFilter;

let context = rclrs::Context::new(std::env::args())?;
let node = Arc::new(rclrs::Node::new(&context, "test")?);
Ros2Logger::new(Arc::clone(&node))
    .with_level(LevelFilter::Off)
    .with_module_level("my_crate", LevelFilter::Info)
    .init()
    .unwrap();

pub fn max_level(&self) -> LevelFilter

Configure the logger

pub fn init(self) -> Result<(), SetLoggerError>

‘Init’ the actual logger and instantiate it, this method MUST be called in order for the logger to be effective.